Remove variables from left-side in config
Which means there are no more variable parts on the left side, which
feels better. This is now transformed when creating the app though to
retain the constant-time lookup of tokens.

This also gives each token a name, which could be used to log accesses for each token once logging gets better.

Closes #5.